dev pin_comment

+++ pinserver/blueprints/pin.py
@@ -83,19 +83,21 @@ def pin_post():
 
 @pin.route('/del_pin/<pin_id>')
 def del_pin(pin_id):
-    if len(pin_id) == 24:
-        pin = Pin.objects(id=pin_id).first()
-        if pin:
-            pin.delete()
+    if g.user_id:
+        if len(pin_id) == 24:
+            pin = Pin.objects(id=pin_id).first()
+            if pin:
+                pin.delete()
 
-            # delete pin from timeline
-            timelines = Timeline.objects(pin=pin_id)
-            for timeline in timelines:
-                timeline.delete()
-                  
-            return jsonify(status='delete success')
-        return jsonify(err_msg='no this pin')
-    return jsonify(err_msg='need pin_id')
+                # delete pin from timeline
+                timelines = Timeline.objects(pin=pin_id)
+                for timeline in timelines:
+                    timeline.delete()
+                      
+                return jsonify(status='delete success')
+            return jsonify(err_msg='no this pin')
+        return jsonify(err_msg='need pin_id')
+    return ('del pin session timeout', 400)
 
 @pin.route('/pins')
 def show_pins():
